refactor: 참가자 행동 열거형 view 패키지로 이동
- 도메인에서 사용되지 않는 열거형을 사용하는 view 패키지로 이동합니다.
