+++ src/main/java/blackjack/view/InputView.java
@@ -1,7 +1,6 @@
 package blackjack.view;
 
 import blackjack.model.participant.Name;
-import blackjack.model.participant.ParticipantAction;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Scanner;
