+++ src/test/java/blackjack/model/participant/ParticipantActionTest.java
@@ -2,6 +2,7 @@ package blackjack.model.participant;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import blackjack.view.ParticipantAction;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
